test(guidance): remove exact wording pins
